requirements.txt
pyTelegramBotAPI
paho-mqtt
mysql-connector-python
dotenv